text formatting for console output, spent too much damn time on this
decided that colored text would be fun, also looking into some reflection to see if it was possible to simplify
the logging functionality. I really didn't want to have to make changes in 5 places, but the more that I tried
to do this, the worse it actually got for managing the functionality. So, I rolled it back and left it.