Lots of get/set methods, also added column cleanup, and column adding functionality
It was pissing me off that there were so many places where I was accessing the class variables
so I changed the calls over to get/set methods
This also helped me to protect some variables, like columns for the worksheet so that I could guarantee they are correct
I also make sure that the access to these is only in one place. This helps keep things organized

The sheets will remove any empty columns and add any missing columns

Downside is they are massively out of order, not sure what to do about that yet.